requirements.txt
fastapi==0.112.2
passlib==1.7.4
pydantic==2.8.2
PyJWT==2.9.0
pytest==8.3.2
uvicorn==0.30.6
python-multipart==0.0.9
